+++ manager/src/Menu/SidebarMenu.php
@@ -33,6 +33,17 @@ class SidebarMenu
 
         $menu->addChild('Work')->setAttribute('class', 'nav-title');
 
+        if ($this->auth->isGranted('ROLE_WORK_MANAGE_PROJECTS')) {
+            $menu->addChild('Projects', ['route' => 'work.projects'])
+                ->setExtra('routes', [
+                    ['route' => 'work.projects'],
+                    ['pattern' => '/^work.projects\..*/']
+                ])
+                ->setExtra('icon', 'nav-icon icon-people')
+                ->setAttribute('class', 'nav-item')
+                ->setLinkAttribute('class', 'nav-link');
+        }
+
         if ($this->auth->isGranted('ROLE_WORK_MANAGE_MEMBERS')) {
             $menu->addChild('Members', ['route' => 'work.members'])
                 ->setExtra('routes', [