+++ manager/src/Model/Work/UseCase/Projects/Project/Create/Command.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Model\Work\UseCase\Projects\Project\Create;
+
+use Symfony\Component\Validator\Constraints as Assert;
+
+class Command
+{
+    /**
+     * @var string
+     * @Assert\NotBlank()
+     */
+    public $name;
+    /**
+     * @var int
+     * @Assert\NotBlank()
+     */
+    public $sort;
+}